Added CLI script GoGoAdhoc.py which can be used to run downloadeds programatically
Additionally, modified the backend to use an asyncio run_until_complete
loop to fix problem of the loop closing when download fails causing an
exception to be raised.

Finally, also added a feature where if the requested download quality is
not found, the highest quality found will be used. This is to avoid
crashes when a particular episode does not have the requested episode
quality.